refactor: query keys to use shared module
