+++ apps/web/src/api/query.keys.ts
@@ -1,37 +1,3 @@
-interface IQueryKeys {
-  currentUnpromotedChanges: string;
-  currentPromotedChanges: string;
-  changesCount: string;
-  myEnvironments: string;
-  currentEnvironment: string;
-  getFeeds: string;
-  getLayoutsList: string;
-  getLayoutById: string;
-  activeIntegrations: string;
-  integrationsList: string;
-  blueprintsList: string;
-  getApiKeys: string;
-  getInAppActive: string;
-  getTemplateById: (templateId?: string) => string;
-  tenantsList: string;
-  getTenantByIdentifier: (tenantIdentifier?: string) => string;
-}
+import { QueryKeys } from '@novu/shared-web';
 
-export const QueryKeys: IQueryKeys = Object.freeze({
-  currentUnpromotedChanges: 'currentUnpromotedChanges',
-  currentPromotedChanges: 'currentPromotedChanges',
-  changesCount: 'changesCount',
-  myEnvironments: 'myEnvironments',
-  currentEnvironment: 'currentEnvironment',
-  getFeeds: 'getFeeds',
-  getLayoutsList: 'getLayoutsList',
-  getLayoutById: 'getLayoutById',
-  activeIntegrations: 'activeIntegrations',
-  integrationsList: 'integrationsList',
-  blueprintsList: 'blueprintsList',
-  getApiKeys: 'getApiKeys',
-  getInAppActive: 'inAppActive',
-  getTemplateById: (templateId?: string) => `notificationById:${templateId}`,
-  tenantsList: 'tenantsList',
-  getTenantByIdentifier: (tenantIdentifier?: string) => `tenantByIdentifier:${tenantIdentifier}`,
-});
+export { QueryKeys };
